create notes under folders + saving text in db

+++ notes.js
@@ -1,7 +1,7 @@
 var express = require('express')
 var router = express.Router()
-var fs = require('fs')
-var mkdirp = require('mkdirp'); 
+// var fs = require('fs')
+// var mkdirp = require('mkdirp'); 
 var showdown = require('showdown');
 converter = new showdown.Converter();
 
@@ -9,19 +9,26 @@ router.get('/', function(req, res) {
     var user = "abc2" // req.session.name
     // var folders = []  
     var notes = [] 
-    req.app.locals.db.collection('notes').find({"user_id":user}).toArray(function(err, docs){
-        if(docs){
-            docs.forEach(function(e,i){
-                notes.push(e.note)
-                console.log(notes)
-            }) 
+
+    // if foldername is null, replace it with "unknown"
+
+    req.app.locals.db.collection('notes').updateMany({user_id:user, folder:null}, {$set:{folder:'unknown'}}, function(err, data){
+        if(err) throw err
+        console.log("null folders -> unknown updated in db!")
+    })
+
+    // user : abc2 ? 
+    // what if no result found? not rendering anything?
+    req.app.locals.db.collection('notes').aggregate([{
+        $group: {
+        _id: "$folder", 
+        notes: { $push: "$note" }
         }
-        
-        // why does res exist here now?
-        console.log("render?")
-        res.render('markdown', {notes:notes})     
-}
-)
+    }]).toArray(function(err, data){
+        console.log(data)
+        res.render('markdown', {fs:data, username:user}) 
+    }
+    )
 })
 
 router.post('/text-to-html', function(req, res){  
@@ -30,39 +37,32 @@ router.post('/text-to-html', function(req, res){
     res.json(html)
 })
 
-router.post('/upsert', function(req, res){
+router.post('/update', function(req, res){
     var user = "abc2" 
     var note = req.body.note
     var content = req.body.content
-    var filename = __dirname + '/' + user + '/' + note + '.txt'
-
-    req.app.locals.db.collection('notes').update({user_id:user, note:note}, {$set:{fpath:filename}}, upsert=true, function(err, data){
+    // text=content in $set - bcoz we search using user-id and note
+    // update() or updateOne()?
+    console.log(note, content)
+    req.app.locals.db.collection('notes').updateOne({user_id:user, note:note}, {$set:{text:content}}, function(err, data){
         if(err) throw err
+        console.log("noted updated in db!", data)
+        res.json({})
     })
     
-    fs.writeFile(filename, content, 'utf8', function(err, data){
-        if(err) throw err
-        res.json({}) 
-    })
 })
 
-// router.post('/edit-folder-name', function(req, res){
-//     var user = "abc2"
-//     var new_name = req.body.new_name
-//     var old_name = req.body.old_name    
-//     var new_filename = __dirname + '/' + user + '/' + new_name 
-//     var old_filename = __dirname + '/' + user + '/' + old_name
-
-//     fs.rename(old_filename, new_filename, function(err, data){  
-//         if(err) throw err             
-//         req.app.locals.db.collection('notes').find({"user_id": user, "folder": old_name})
-//             .forEach(function(e, i){
-//                 e.folder = new_name,
-//                 e.fpath = new_filename + "/" + e.note + ".txt";
-//                 req.app.locals.db.collection('notes').save(e)
-//             })
-//     })
-// })
+router.post('/edit-folder-name', function(req, res){
+    var user = "abc2"
+    var new_name = req.body.new_name
+    var old_name = req.body.old_name          
+    req.app.locals.db.collection('notes').find({"user_id": user, "folder": old_name})
+            .forEach(function(e, i){
+                e.folder = new_name,
+                req.app.locals.db.collection('notes').save(e)
+                console.log("updated folder name in db", "from -", old_name, "to -", new_name)
+        })
+    })
 
 
 router.post('/edit-note-name', function(req, res){
@@ -72,66 +72,55 @@ router.post('/edit-note-name', function(req, res){
     req.app.locals.db.collection('notes').find({"user_id": user, "note": old_name})
         .forEach(function(e, i){
             e.note = new_name,
-            old_path = e.fpath
-            e.fpath = __dirname + "/" + user + "/" + new_name + ".txt";
             req.app.locals.db.collection('notes').save(e)
-
-            fs.rename(old_path, e.fpath, function(err, data){
-                if(err) throw err
+            console.log("updated note name in db", "from -", old_name, "to -", new_name)
         })
     })
-
-})
 
 
 router.get('/read', function(req, res){
     var user = "abc2"
     var folder = req.query.folder
     var note = req.query.note
-    var filename = __dirname + '/' + user
-
-    fs.readFile(filename + '/' + note + '.txt', 'utf8', function(err, data){
-        if(err){
-            mkdirp(filename, function(err, data){
-                if (err) throw err
-                // extra space in #pad bcoz i had " " as content
-                fs.writeFile(filename + '/' + note + '.txt', "", 'utf8', function(err, data){
-                        if(err) throw err
-                
-                        req.app.locals.db.collection('notes').insert({user_id:user, note:note, fpath:filename + '/' + note + '.txt'}, function(err, data){
-                            if(err) throw err
-                            console.log("inserted for the 1st time in db....")
-                        })
-
-                        res.send("")
-
-                    })
-            });    
+    console.log("{ user_id:",user, ", note:",note, " ,folder:", folder, " }")
+    req.app.locals.db.collection('notes').find({user_id:user, note:note, folder:folder}).toArray(function(err, docs){
+        if(err) throw err
+        if(docs.length){
+            console.log(docs[0].text, "doc found, thus sending the stored text")
+            var text = docs[0].text
+            res.send(text)
         }
         else{
-            var content = data
-            res.send(content)
-        }        
+            console.log(docs, "not found, thus inserting")
+            req.app.locals.db.collection('notes').insert({user_id:user, note:note, folder:folder, text:""}, function(err, data){
+                if(err) throw err
+                console.log("inserted for the 1st time in db....")
+                res.send("")
+            })
+        }
     })
 })
+
 
 router.post('/delete-a-note', function(req, res){
     var user = "abc2"
     var note = req.body.note
-    var path = '' 
-    req.app.locals.db.collection('notes').find({"user_id":user, "note":note}).forEach(function(e,i){
-        path = e.fpath
-        fs.unlink(path, function(err, data){
-            if(err) throw err
-            console.log("note deleted from file-sys")
-        })
-        req.app.locals.db.collection('notes').deleteOne({"note":note}, function(err, result){
+    var folder = req.body.folder
+    req.app.locals.db.collection('notes').deleteOne({"user_id":user, "note":note, folder:folder}, function(err, result){
             if(err) throw err
             console.log("note deleted from db")
             }
         )
     })     
-})
 
+router.post('/delete-a-folder', function(req, res){
+    var user = "abc2"
+    var folder = req.body.folder
+    req.app.locals.db.collection('notes').deleteOne({"user_id":user, "folder":folder}, function(err, result){
+            if(err) throw err
+            console.log("folder and corresponding notes deleted from db")
+            }
+        )
+    }) 
 
 module.exports = router
